feat: consolidate platform implementations by deleting `file_picker_io.dart` and unifying `MethodChannelFilePicker` with platform-specific codecs and updated method signatures.
