+++ lib/file_picker.dart
@@ -1,7 +1,7 @@
 export 'src/file_picker.dart';
 export 'src/platform_file.dart';
 export 'src/file_picker_result.dart';
-export 'src/file_picker_io.dart';
+export 'src/file_picker_method_channel.dart';
 export 'src/file_picker_macos.dart';
 export 'src/linux/file_picker_linux.dart';
 export 'src/windows/file_picker_windows.dart';
